# generate/openapi.json
{"openapi":"3.0.0","paths":{"/auth/signup":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/check":{"post":{"operationId":"UserController_findEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDuplicateDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"UserController_loginUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/note":{"get":{"operationId":"NoteController_fetchAllNote","parameters":[],"responses":{"200":{"description":""}},"tags":["Note"]}},"/note/{id}":{"get":{"operationId":"NoteController_fetchNoteById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Note"]},"delete":{"operationId":"NoteController_deleteNoteById","summary":"Check to see if an authentication token is still valid.","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Note"],"security":[{"bearer":[]}]}},"/note/create":{"post":{"operationId":"NoteController_createNote","summary":"Check to see if an authentication token is still valid.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNoteDto"}}}},"responses":{"201":{"description":""}},"tags":["Note"],"security":[{"bearer":[]}]}}},"info":{"title":"Diary API","description":"The Diary API Documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"}},"required":["email","name","password"]},"EmailDuplicateDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"LoginUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateNoteDto":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"}},"required":["title","content"]}}}}